+++ src/main/java/com/saditec/platform/controller/EducationalAndRecreationalFacilitiesController.java
@@ -2,10 +2,13 @@ package com.saditec.platform.controller;
 
 import com.saditec.platform.entity.TEducationalAndRecreationalFacilityEntity;
 import com.saditec.platform.service.EducationalAndRecreationalFacilitiesService;
+import com.saditec.platform.type.ApiResponse;
+import com.saditec.platform.type.TEducationalAndRecreationalFacilityTypeDto;
 import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
+
+import java.security.Principal;
+import java.util.List;
 
 @RestController
 @RequestMapping("/educational/recreational/facilities")
@@ -15,7 +18,20 @@ public class EducationalAndRecreationalFacilitiesController {
     private final EducationalAndRecreationalFacilitiesService educationalAndRecreationalFacilitiesService;
 
     @GetMapping
-    public TEducationalAndRecreationalFacilityEntity add(){
+    public TEducationalAndRecreationalFacilityEntity add() {
         return educationalAndRecreationalFacilitiesService.add();
+    }
+
+    @GetMapping("/type")
+    public ApiResponse<List<TEducationalAndRecreationalFacilityTypeDto>> getAllEducationalAndRecreationalFacility() {
+        return new ApiResponse<List<TEducationalAndRecreationalFacilityTypeDto>>()
+                .toSuccess(educationalAndRecreationalFacilitiesService.getAllEducationalAndRecreationalFacility());
+    }
+
+    @PostMapping("/type")
+    public ApiResponse<TEducationalAndRecreationalFacilityTypeDto> addEducationalAndRecreationalFacilityType(@RequestBody TEducationalAndRecreationalFacilityTypeDto tEducationalAndRecreationalFacilityTypeDto,
+                                                                                                             Principal principal) {
+        return new ApiResponse<TEducationalAndRecreationalFacilityTypeDto>()
+                .toSuccess(educationalAndRecreationalFacilitiesService.addEducationalAndRecreationalFacilityType(tEducationalAndRecreationalFacilityTypeDto, principal));
     }
 }