+++ src/main/java/com/saditec/platform/entity/TEducationalAndRecreationalFacilityTypeEntity.java
@@ -1,10 +1,13 @@
 package com.saditec.platform.entity;
 
+import com.saditec.platform.type.TEducationalAndRecreationalFacilityTypeDto;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.time.Instant;
 import java.util.Set;
+import java.util.UUID;
 
 @Getter
 @Setter
@@ -26,4 +29,31 @@ public class TEducationalAndRecreationalFacilityTypeEntity extends TBaseEntity {
 
     @OneToMany(mappedBy = "type")
     private Set<TEducationalAndRecreationalFacilityEntity> educationalAndRecreationalFacilities;
+
+    public TEducationalAndRecreationalFacilityTypeDto toDto() {
+        return TEducationalAndRecreationalFacilityTypeDto.builder()
+                .identifier(getIdentifier())
+                .name(getName())
+                .description(getDescription())
+                .image(getImage())
+                .imagePublic(getImagePublic())
+                .build();
+    }
+
+    public static TEducationalAndRecreationalFacilityTypeEntity toEntity(TEducationalAndRecreationalFacilityTypeDto tEducationalAndRecreationalFacilityTypeDto, String user) {
+
+        TEducationalAndRecreationalFacilityTypeEntity tEducationalAndRecreationalFacilityTypeEntity = new TEducationalAndRecreationalFacilityTypeEntity();
+        tEducationalAndRecreationalFacilityTypeEntity.setIdentifier(UUID.randomUUID().toString());
+        tEducationalAndRecreationalFacilityTypeEntity.setName(tEducationalAndRecreationalFacilityTypeDto.getName());
+        tEducationalAndRecreationalFacilityTypeEntity.setDescription(tEducationalAndRecreationalFacilityTypeDto.getDescription());
+        tEducationalAndRecreationalFacilityTypeEntity.setImage(tEducationalAndRecreationalFacilityTypeDto.getImage());
+        tEducationalAndRecreationalFacilityTypeEntity.setImagePublic(tEducationalAndRecreationalFacilityTypeDto.getImagePublic());
+        tEducationalAndRecreationalFacilityTypeEntity.setEnabled(true);
+        tEducationalAndRecreationalFacilityTypeEntity.setCreatedBy(user);
+        tEducationalAndRecreationalFacilityTypeEntity.setCreatedAt(Instant.now());
+        tEducationalAndRecreationalFacilityTypeEntity.setUpdatedBy(user);
+        tEducationalAndRecreationalFacilityTypeEntity.setUpdatedAt(Instant.now());
+
+        return tEducationalAndRecreationalFacilityTypeEntity;
+    }
 }