+++ src/main/java/com/saditec/platform/repository/IEducationalAndRecreationalFacilitiesTypeRepository.java
@@ -1,0 +1,9 @@
+package com.saditec.platform.repository;
+
+import com.saditec.platform.entity.TEducationalAndRecreationalFacilityTypeEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface IEducationalAndRecreationalFacilitiesTypeRepository extends JpaRepository<TEducationalAndRecreationalFacilityTypeEntity, String> {
+}